Get physiologies

getPhysiologies get all physiology annotations for one or more taxa.

getPhysiologies(x, phys = physiologies(), taxid.type = "Taxon_name")

Arguments

x

A valid NCBI ID or taxon name

phys

A list of dataframes with physiologies imported with the physiologies function. If not specified, all physiologies will be imported.

taxid.type

The type of taxid. Valid options: Taxon_name or NCBI_ID. Default is Taxon_name.

Value

A list of physiologies per taxa.

Examples

phys <- physiologies()
#> Finished acetate producing
#> Finished aerophilicity
#> Finished animal pathogen
#> Finished antimicrobial resistance
#> Dropped 3 rows with missing Attribute_value from antimicrobial sensitivity
#> Finished arrangement
#> Finished biofilm forming
#> Dropped 4 rows with missing Attribute_value from butyrate producing
#> Finished COGEM pathogenicity rating
#> Finished disease association
#> Finished extreme environment
#> Finished gram stain
#> Finished growth medium
#> Finished growth temperature
#> Finished habitat
#> Finished health associated
#> Dropped 10 rows with missing Attribute_value from hydrogen gas producing
#> Finished isolation site
#> Dropped 9 rows with missing Attribute_value from lactate producing
#> Finished length
#> Finished mutation rate per site per generation
#> Finished mutation rates per site per year
#> Finished optimal ph
#> Finished plant pathogenicity
#> Finished shape
#> Finished spore shape
#> Finished width
#> Dropped 5 rows with missing Attribute_value from genome size
#> Finished coding genes
#> Finished sphingolipid producing

taxa_names <- c('Escherichia coli', 'Staphylococcus aureus')
taxa_ids <- c('562', '1280')

phys_names <- getPhysiologies(taxa_names, phys)
phys_ids <- getPhysiologies(taxa_ids, phys, 'NCBI_ID')