Get signatures

getSignatures creates a list of signatures from a bugphyzz data frame.

getSignatures(
  df,
  tax.id.type = "Taxon_name",
  tax.level = "mixed",
  Frequency = c("unknown", "sometimes", "usually", "always"),
  Evidence = c("asr", "unknown", "igc", "asr"),
  min.size = 1,
  min = NULL,
  max = NULL
)

Arguments

df

A data frame from bugphyzz.

tax.id.type

Type of taxids. Check types with the whichID function.

tax.level

A character vector with taxonomic ranks, e.g., genus, species, strain.

Frequency

A character vector with one or more values of: unknown, sometimes, usually, always.

Evidence

A character vector with one or more values of: unknown, sometimes, usually, always.

min.size

The minimun number of elements in a signatures. Default is 1.

min

Minimum threshold for attributes with continuous values. It's ignored when creating signatures for categorical attributes.

max

Maximum threshold for attributes with continuous values. It's ignored when creating signatures for categorical attributes.

Value

A list of signatures. NULL if no signatures.

Examples

aer <- physiologies('aerophilicity')[[1]]
#> Finished aerophilicity
aer_sig <- getSignatures(aer)
#> Warning: Dropping 3648 rows with FALSE values.
#> Retrieving all levels of taxa.
#> Creating signature for categorical attributes.

ph <- physiologies('optimal ph')[[1]]
#> Finished optimal ph
ph_sig <- getSignatures(ph)
#> Retrieving all levels of taxa.
#> Creating signature for continuous attribute.