Import bugphyzz

importBugphyzz imports bugphyzz annotations as a list of tidy data.frames. To learn more about the structure of the data.frames please check the bugphyzz vignette with browseVignettes("bugphyzz").

importBugphyzz(
  version = "devel",
  force_download = FALSE,
  v = 0.5,
  exclude_rarely = TRUE
)

Arguments

version

Character string indicating the version. Options: devel, doi, GitHub hash.

force_download

Logical value. Force a fresh download of the data or use the one stored in the cache (if available). Default is FALSE.

v

Validation value. Default 0.5 (see details).

exclude_rarely

Default is TRUE. Exclude values with Frequency == FALSE (see details).

Value

A list of tidy data frames.

Details

Data structure

The data structure of the data.frames imported with importBugphyzz are detailed in the main vignette. Please run browseVignettes("bugphyzz").

Validation (v argument)

Data imported with importBugphyzz includes annotations imputed through ancestral state reconstruction (ASR) methods. A 10-fold cross-validation approach was implemented to assess the reliability of the data imputed. Mathew's correlation coefficient (MCC) and R-squared (R2) were used for the validation of discrete and numeric attributes. Details can be found at: https://github.com/waldronlab/taxPProValidation. By default, imputed annotations with a MCC or R2 value greater than 0.5 are imported. The minimum value can be adjusted with the v argument (only values between 0 and 1).

Frequency (exclude_rarely argument)

One of the variables in the bugphyzz data.frames is "Frequency", which can adopt values of "always", "usually", "sometimes", "rarely", or "never". By default "never" and "rarely" are excluded. "rarely" could be included with exclude_rarely = FALSE. To learn more about these frequency keywords please check the bugphyzz vignette with browseVignettes("bugphyzz").

Examples

bp <- importBugphyzz()
#> Importing multistate data...
#> Using data downloaded on 2024-03-22 18:19:07.
#> Importing binary data...
#> Using data downloaded on 2024-03-22 18:19:16.
#> Importing numeric data...
#> Using data downloaded on 2024-03-22 18:19:19.
names(bp)
#>  [1] "aerophilicity"                        
#>  [2] "antimicrobial resistance"             
#>  [3] "arrangement"                          
#>  [4] "biosafety level"                      
#>  [5] "cogem pathogenicity rating"           
#>  [6] "disease association"                  
#>  [7] "gram stain"                           
#>  [8] "habitat"                              
#>  [9] "hemolysis"                            
#> [10] "shape"                                
#> [11] "spore shape"                          
#> [12] "animal pathogen"                      
#> [13] "antimicrobial sensitivity"            
#> [14] "biofilm formation"                    
#> [15] "butyrate-producing bacteria"          
#> [16] "extreme environment"                  
#> [17] "health associated"                    
#> [18] "host-associated"                      
#> [19] "hydrogen gas producing"               
#> [20] "lactate producing"                    
#> [21] "motility"                             
#> [22] "plant pathogenicity"                  
#> [23] "sphingolipid producing"               
#> [24] "spore formation"                      
#> [25] "coding genes"                         
#> [26] "genome size"                          
#> [27] "growth temperature"                   
#> [28] "length"                               
#> [29] "mutation rate per site per generation"
#> [30] "mutation rate per site per year"      
#> [31] "optimal ph"                           
#> [32] "width"