Import bugphyzz (numeric/continuous data)

importBugphyzzNumeric imports a list of data.frames containing numeric attributes. In this case, the attributes are in separate data.frames because they can have different units or scales (e.g. Celsius degrees, ph, etc.)

importBugphyzzNumeric(
  keyword = "all",
  version = "devel",
  force_download = FALSE
)

Arguments

version

Character string. The version to download. Default is 'devel' (current file on the GitHub repo waldronlab/bugphyzzExports).

force_download

Logical value. Force a fresh download of the data or use the one stored in the cache (if available). Default is FALSE.

Value

A list of data.frames.

Examples

bp_num <- importBugphyzzNumeric()
#> Using data downloaded on 2023-04-04 17:56:21.

## Check available numeric attributes
names(bp_num)
#> [1] "growth temperature" "optimal ph"        

## Select growth temperature
gt <- bp_num[['growth temperature']]

## Get taxa that grows better betwen 0 and 25 Celsius degrees
sub_gt <- gt[which(gt$Attribute_value_min >= 0 & gt$Attribute_value_max <= 25),]

## Creat signature at the genus level
sigs <- getBugphyzzSignatures(sub_gt, tax.id.type = 'Taxon_name', tax.level = 'genus')
head(sigs[[1]])
#> [1] "Natronobiforma"        "Saliphagus"            "Bryocella"            
#> [4] "Candidatus Koribacter" "Occallatibacter"       "Pseudacidobacterium"