Make signatures

makeSignatures creates a list of microbe signatures from a bugphyzz dataset, i.e. imported with either the physiologies or the fattyAcidComposition functions.

makeSignatures(df, taxids, tax_rank, min_sig_size = 5, inherited = FALSE)

Arguments

df
A bugphyzz dataset.

taxids
A character string indicating the type of taxid. Valid options: NCBI_ID, Taxon_name, Genome_ID, Accession_number.

tax_rank
A character string indicating the taxonomic rank of the microbes in the signatures. Valid options: superkingdom, kingdom, phylum, class, order, family, species, strain.

min_sig_size
The minimum number of microbes in a signature.

inherited
If TRUE inherited signatures are created. Default is FALSE.

Value

A named list of microbe signatures.

Inherited signatures

makeSignatures allows the creation of inherited signatures when the inherited argument is set to TRUE. This means that a taxon of a given taxonomic rank (e.g. species) can inherit an attribute annotation from a taxon of a lower rank (e.g. strain) if they belong to the same lineage. For example, if the Pyrococcus abyssi species is not annotated in a given bugphyzz dataset, but the Pyrococcus furiosus DSM 3638 strain is, then the inherited = TRUE argument allows to create a signature at the species level including the Pyrococcus abyssi species.

Examples

library(bugphyzz)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

## Make signatures of thermophilic microbes

gt <- physiologies("growth temperature")[[1]]

#> Finished growth temperature

gt_therm <- gt %>%
    filter(Attribute_value > 80)

gt_sig <- makeSignatures(gt_therm, taxids = "NCBI_ID", tax_rank = "species")
gt_sig

#> $`growth temperature`
#>  [1] "644281"  "2714"    "33934"   "63363"   "28095"   "498846"  "67760"  
#>  [8] "2320"    "2180"    "29292"   "1008460" "1295609" "2269"    "1200300"
#> [15] "37636"  
#> 

## Make inherited signatures

gt_sig_inherired <- makeSignatures(gt_therm, taxids = "NCBI_ID",
                                   tax_rank = "species", inherited = TRUE)
gt_sig_inherired

#> $`growth temperature`
#>  [1] "56636"   "63363"   "2234"    "76887"   "94694"   "54261"   "54248"  
#>  [8] "334771"  "67760"   "2190"    "644281"  "2320"    "13773"   "181486" 
#> [15] "2277"    "29292"   "2261"    "53953"   "84599"   "2280"    "2287"   
#> [22] "187878"  "311400"  "2269"    "70771"   "164451"  "2714"    "33934"  
#> [29] "28095"   "498846"  "2180"    "1008460" "1295609" "1200300" "37636"  
#>