Display number of taxa without NCBI ID per attribute in physiology dataset

Display number of taxa without NCBI ID per attribute in physiology dataset

num_taxa_without_ncbi_per_attribute(data, phys)

Arguments

data
a tibble object, converted from the output by the physiologies function and has been converted by the fetch_ranks function

phys
a string vector of the physiology name

Value

a ggplot

Examples

physiologies <- physiologies()

#> Finished animal pathogen

#> Finished antimicrobial resistance

#> Dropped 3054 rows with missing Attribute_value from antimicrobial sensitivity

#> Finished biofilm forming

#> Dropped 4 rows with missing Attribute_value from butyrate producing

#> Finished acetate producing

#> Finished lactate producing

#> Finished arrangement

#> Dropped 9304 rows with missing Attribute_value from shape

#> Finished COGEM pathogenicity rating

#> Finished mutation rate per site per generation

#> Finished mutation rates per site per year

#> Finished extreme environment

#> Finished gram stain

#> Finished growth medium

#> Finished growth temperature

#> Finished habitat

#> Finished optimal ph

#> Finished aerophilicity

#> Finished plant pathogenicity

#> Finished width

#> Dropped 1 rows with missing Attribute_value from spore shape

#> Finished isolation site

#> Finished disease association

#> Finished hydrogen gas producing

#> Finished length

#> Finished health associated

aer <- physiologies[["aerophilicity"]] %>%
as_tibble()
aer_ranks <- fetch_ranks(aer)

#> Error in fetch_ranks(aer): could not find function "fetch_ranks"

num_taxa_without_ncbi_per_attribute(aer_ranks, "aerophilicity")

#> Error in num_taxa_without_ncbi_per_attribute(aer_ranks, "aerophilicity"): could not find function "num_taxa_without_ncbi_per_attribute"