Import phsiologies

physiologies imports data from the Google spreadsheets at https://drive.google.com/drive/folders/1i2UAolVWAYa7UnETNnCs0BDWjKPp3ev5. This function (and its internal functions) do minimal changes to the imported data. These changes are only meant to match data coming from different sources, and attaching information needed for further processing, such as source and attribute type.

physiologies(keyword = "all", full_source = FALSE)

Arguments

keyword

Character vector with one or more valid keywords. Valid keyboards can be checked with showPhys. If 'all', all physiologies are imported.

full_source

Logical. If TRUE, the Attribute_source column will contain full source information. If FALSE, the Attribute_source column will contain shortened versions of the sources. Default is FALSE.

Value

A list of data.frames in tidy format.

Examples

l <- physiologies('all')
#> All physiologies will be imported.
#> Finished COGEM pathogenicity rating.
#> Finished acetate producing.
#> Finished aerophilicity.
#> Finished animal pathogen.
#> Finished antimicrobial resistance.
#> Finished antimicrobial sensitivity.
#> Finished arrangement.
#> Finished biofilm forming.
#> Finished biosafety level.
#> Finished butyrate producing.
#> Finished coding genes.
#> Finished country.
#> Finished disease association.
#> Finished extreme environment.
#> Finished genome size.
#> Finished geographic location.
#> Finished gram stain.
#> Finished growth medium.
#> Finished growth temperature.
#> Finished habitat.
#> Finished halophily.
#> Finished health associated.
#> Finished hemolysis.
#> Finished host-associated.
#> Finished hydrogen gas producing.
#> Finished isolation site.
#> Finished lactate producing.
#> Finished length.
#> Finished metabolite production.
#> Finished metabolite utilization.
#> Finished motility.
#> Finished mutation rate per site per generation.
#> Finished mutation rate per site per year.
#> Finished optimal ph.
#> Finished pathogenicity human.
#> Finished plant pathogenicity.
#> Finished shape.
#> Finished sphingolipid producing.
#> Finished spore formation.
#> Finished spore shape.
#> Finished width.
#> Warning: Missing columns in halophily. Missing columns are: Genome_ID, Accession_ID
df <- physiologies('aerophilicity')[[1]]
#> Finished aerophilicity.