Which Attributes

whichAttr shows which attributes are present in a dataset imported with importBugphyzz. This would be the names of the singatures created with getBugphyzzSignatures.

whichAttr(bp)

Arguments

bp

A data.frame imported with importBugphyzz.

Value

A character vector.

Examples

bp <- importBugphyzz()
#> Using data downloaded on 2023-07-26 17:35:36.
whichAttr(bp)
#>   [1] "10% salts"                              
#>   [2] "20% salts"                              
#>   [3] "30% salts"                              
#>   [4] "5% salts"                               
#>   [5] "5% sheep blood agar"                    
#>   [6] "Acne"                                   
#>   [7] "Arthritis"                              
#>   [8] "Blunervirus"                            
#>   [9] "Burkholderia cepacia"                   
#>  [10] "Butyrate-Producing Bacteria--TRUE"      
#>  [11] "CAMP test"                              
#>  [12] "COGEM pathogenicity rating 1"           
#>  [13] "COGEM pathogenicity rating 2"           
#>  [14] "COGEM pathogenicity rating 3"           
#>  [15] "Caries"                                 
#>  [16] "Conjunctivitis"                         
#>  [17] "DSMZ Medium 1000"                       
#>  [18] "Extreme environment--FALSE"             
#>  [19] "Extreme environment--TRUE"              
#>  [20] "Flavobacterium psychrophilum"           
#>  [21] "Foodborne Illness"                      
#>  [22] "Furunculosis"                           
#>  [23] "Gangrene"                               
#>  [24] "Gastroenteritis"                        
#>  [25] "Glanders"                               
#>  [26] "Gonorrhea"                              
#>  [27] "Haemophilus"                            
#>  [28] "Health associated--FALSE"               
#>  [29] "Health associated--TRUE"                
#>  [30] "Infertility"                            
#>  [31] "Inflammation"                           
#>  [32] "Lyme Disease"                           
#>  [33] "Microcystis aeruginosa"                 
#>  [34] "Mycobacterial Infection"                
#>  [35] "Periodontal disorder"                   
#>  [36] "Plant chlorosis"                        
#>  [37] "Reaction to Leaf spot"                  
#>  [38] "Sphingolipid producing--TRUE"           
#>  [39] "Terrabacteria group"                    
#>  [40] "Toxic Shock Syndrome"                   
#>  [41] "Tuberculosis"                           
#>  [42] "Tumor"                                  
#>  [43] "ZoBell marine agar"                     
#>  [44] "abortion"                               
#>  [45] "abscess"                                
#>  [46] "acetate producing--TRUE"                
#>  [47] "acidic"                                 
#>  [48] "acrodermatitis"                         
#>  [49] "aerobic"                                
#>  [50] "aerotolerant"                           
#>  [51] "african tick-bite fever"                
#>  [52] "alkaline"                               
#>  [53] "alpha"                                  
#>  [54] "anaerobic"                              
#>  [55] "anaplasmosis"                           
#>  [56] "anemia"                                 
#>  [57] "animal pathogen--FALSE"                 
#>  [58] "animal pathogen--TRUE"                  
#>  [59] "anthrax"                                
#>  [60] "antimicrobial sensitivity--FALSE"       
#>  [61] "antimicrobial sensitivity--TRUE"        
#>  [62] "artificial seawater"                    
#>  [63] "asp-2"                                  
#>  [64] "atypical pneumonia"                     
#>  [65] "average"                                
#>  [66] "bacillary angiomatosis"                 
#>  [67] "bacillus"                               
#>  [68] "bacteremia"                             
#>  [69] "bacterial leaf blight"                  
#>  [70] "bacterial vaginosis"                    
#>  [71] "bartonellosis"                          
#>  [72] "bent filament"                          
#>  [73] "beta"                                   
#>  [74] "bg11"                                   
#>  [75] "biofilm formation--FALSE"               
#>  [76] "biofilm formation--TRUE"                
#>  [77] "biosafety level 1"                      
#>  [78] "biosafety level 2"                      
#>  [79] "biosafety level 3"                      
#>  [80] "blood infection"                        
#>  [81] "bordetellosis"                          
#>  [82] "botulism"                               
#>  [83] "brackish microbiological culture medium"
#>  [84] "branched"                               
#>  [85] "brucellosis"                            
#>  [86] "c"                                      
#>  [87] "canker"                                 
#>  [88] "carbapenem-resistant Acinetobacter"     
#>  [89] "casitone-yeast extract agar"            
#>  [90] "cat scratch fever"                      
#>  [91] "cell chain"                             
#>  [92] "cell cluster"                           
#>  [93] "cell lysis"                             
#>  [94] "charcoal-yeast extract"                 
#>  [95] "chocolate agar"                         
#>  [96] "cholera"                                
#>  [97] "chronic granulomatous disease"          
#>  [98] "clavate"                                
#>  [99] "coccobacillus"                          
#> [100] "coccus"                                 
#> [101] "colitis"                                
#> [102] "conway"                                 
#> [103] "crown gall"                             
#> [104] "cystitis"                               
#> [105] "defined microbiological culture medium" 
#> [106] "diarrhea"                               
#> [107] "diphtheria"                             
#> [108] "discoid"                                
#> [109] "dumbbell-shaped"                        
#> [110] "duodenal lesions"                       
#> [111] "dysentery"                              
#> [112] "ehrlichiosis"                           
#> [113] "elliptic"                               
#> [114] "encephalitis"                           
#> [115] "endocarditis"                           
#> [116] "endospore"                              
#> [117] "erysipelas"                             
#> [118] "facultatively anaerobic"                
#> [119] "fast"                                   
#> [120] "fever"                                  
#> [121] "filamentous"                            
#> [122] "fildes enrichment"                      
#> [123] "flat"                                   
#> [124] "gamma"                                  
#> [125] "gastroenteritis"                        
#> [126] "gram stain negative"                    
#> [127] "gram stain positive"                    
#> [128] "gram stain variable"                    
#> [129] "gse"                                    
#> [130] "heart infusion agar"                    
#> [131] "heartwater"                             
#> [132] "human granulocytic anaplasmosis"        
#> [133] "hydrogen gas producing--FALSE"          
#> [134] "hydrogen gas producing--TRUE"           
#> [135] "hyperthermophile"                       
#> [136] "infection"                              
#> [137] "kidney disease"                         
#> [138] "lactate producing--FALSE"               
#> [139] "lactate producing--TRUE"                
#> [140] "large"                                  
#> [141] "lb agar"                                
#> [142] "lb medium"                              
#> [143] "legionnaire's disease"                  
#> [144] "leprosy"                                
#> [145] "mastitis"                               
#> [146] "medium"                                 
#> [147] "medium 31 agar"                         
#> [148] "medium fast"                            
#> [149] "melioidosis"                            
#> [150] "meningitis"                             
#> [151] "meningoencephalitis"                    
#> [152] "mesophile"                              
#> [153] "modified shieh agar"                    
#> [154] "motility--FALSE"                        
#> [155] "motility--TRUE"                         
#> [156] "mrs broth"                              
#> [157] "mueller-hinton"                         
#> [158] "neutral"                                
#> [159] "not resistant to beta-lactam"           
#> [160] "nutrient broth"                         
#> [161] "paired cells"                           
#> [162] "palisade"                               
#> [163] "pathogenicity human--TRUE"              
#> [164] "plague"                                 
#> [165] "plant pathogenity--FALSE"               
#> [166] "plant pathogenity--TRUE"                
#> [167] "pleiomorphic cell"                      
#> [168] "pneumonia"                              
#> [169] "polyarthritis"                          
#> [170] "polyhedral shape"                       
#> [171] "psychrophile"                           
#> [172] "pyelonephritis"                         
#> [173] "pyg medium"                             
#> [174] "reaction to black rot"                  
#> [175] "rectangle"                              
#> [176] "resistance to erythromycin A"           
#> [177] "resistance to isoniazid"                
#> [178] "resistance to levofloxacin"             
#> [179] "resistance to methicillin"              
#> [180] "resistance to penicillin"               
#> [181] "resistance to rifampicin"               
#> [182] "resistance to streptomycin"             
#> [183] "resistance to tetracycline"             
#> [184] "resistance to trimethoprim"             
#> [185] "resistant to beta-lactam"               
#> [186] "respiratory infection"                  
#> [187] "respiratory system disease"             
#> [188] "reverse Y-branched filaments"           
#> [189] "rosettes"                               
#> [190] "sensitive to amikacin"                  
#> [191] "sensitive to capreomycin"               
#> [192] "sensitive to carbapenems"               
#> [193] "sensitive to chloramphenicol"           
#> [194] "sensitive to ciprofloxacin"             
#> [195] "sensitive to clindamycin"               
#> [196] "sensitive to ertapenem"                 
#> [197] "sensitive to erythromycin A"            
#> [198] "sensitive to ethambutol"                
#> [199] "sensitive to gentamicin"                
#> [200] "sensitive to isoniazid"                 
#> [201] "sensitive to kanamycin A"               
#> [202] "sensitive to levofloxacin"              
#> [203] "sensitive to meropenem"                 
#> [204] "sensitive to methicillin"               
#> [205] "sensitive to ofloxacin"                 
#> [206] "sensitive to rifampicin"                
#> [207] "sensitive to streptomycin"              
#> [208] "sensitive to tetracycline"              
#> [209] "sensitive to trimethoprim"              
#> [210] "septicemia"                             
#> [211] "sigmoid"                                
#> [212] "single"                                 
#> [213] "skin irritation"                        
#> [214] "slow"                                   
#> [215] "sm agar"                                
#> [216] "small"                                  
#> [217] "sorghum"                                
#> [218] "spirillum"                              
#> [219] "spore formation--FALSE"                 
#> [220] "spore formation--TRUE"                  
#> [221] "staley pyg"                             
#> [222] "staphylococcus"                         
#> [223] "syphilis"                               
#> [224] "tail"                                   
#> [225] "tetrads"                                
#> [226] "tetrahedral"                            
#> [227] "thermophile"                            
#> [228] "toxin"                                  
#> [229] "trench fever"                           
#> [230] "triangle"                               
#> [231] "tryptic soy agar"                       
#> [232] "tryptic soy broth"                      
#> [233] "tuberculosis"                           
#> [234] "tumors"                                 
#> [235] "typhoid"                                
#> [236] "typhus"                                 
#> [237] "u-shaped"                               
#> [238] "ulcer"                                  
#> [239] "unbranched"                             
#> [240] "urinary tract infection"                
#> [241] "v-formation"                            
#> [242] "very alkaline"                          
#> [243] "very large"                             
#> [244] "very small"                             
#> [245] "wrinkled"