Make signatures

makeSignatures creates a list of microbe signatures from a bugphyzz dataset, i.e., a dataset imported either with the physiologies or the fattyAcidComposition functions.

makeSignatures(
  df,
  taxids,
  sig_type = "regular",
  tax_rank = NULL,
  min_sig_size = 1
)

Arguments

df
A bugphyzz dataset.

taxids
A character string indicating the type of taxid. Valid options: NCBI_ID, Taxon_name, Genome_ID, Accession_number.

tax_rank
A character string indicating the taxonomic rank of the microbes in the signatures. Valid options: superkingdom, kingdom, phylum, class, order, family, species, strain.

min_sig_size
The minimum number of microbes in a signature.

type
A character string. Valid options: regular, normal, inherited.

Value

A named list of microbe signatures.

Types of signatures

makeSignatures allows the creation of three types of signatures (regular, ranked, or inherited) with the type argument.

The regular option (default) creates a list of signatures per attribute considering all of the taxids and attributes in the dataset, regardless of taxonomic rank.

The ranked option filters the taxids by taxonomic rank before creating the lists of signatures. Currently, only "species", "genus", and "strain" rank levels are supported.

The inherited option allows the creation of signatures in which the members of the list inherit an attribute annotation from a taxon of a lower level, as long as they belong to the same lineage. For example, if the Pyrococcus abyssi species is not annotated in a given bugphyzz dataset, but the Pyrococcus furiosus DSM 3638 strain is, then the type = "inherited" argument allows to create a signature at the species level including the Pyrococcus abyssi species.

Examples

library(bugphyzz)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

## Example 1- Crete a list of signatures from the aerophilicity dataset
aer <- physiologies("aerophilicity")[[1]]

#> Finished aerophilicity

aer_sig <- makeSignatures(aer, taxids = "NCBI_ID", min_sig_size = 5)

## Exampl 2 - Create signatures of thermophilic bacteria
gt <- physiologies("growth temperature")[[1]]

#> Finished growth temperature

tm <- gt |> filter(Attribute_value > 80)

## Species level signatures with ranked option
tm_sig_ranked <- makeSignatures(tm, taxids = "Taxon_name",
                                tax_rank = "species", sig_type = "ranked")

## Species level signtures with inherited option
tm_sig_inherited <- makeSignatures(tm, taxids = "Taxon_name",
                                   tax_rank = "species",
                                   sig_type = "inherited")

## Compare lengths
length(tm_sig_inherited[[1]]) > length(tm_sig_ranked[[1]])

#> [1] TRUE