Which Attributes

whichAttr shows which attributes are present in a dataset imported with importBugphyzz. This would be the names of the singatures created with getBugphyzzSignatures.

whichAttr(bp)

Arguments

bp

A data.frame imported with importBugphyzz.

Value

A character vector.

Examples

bp <- importBugphyzz()
#> Using data downloaded on 2023-07-17 18:53:43.
#> Warning: The following named parsers don't match the column names: PATRIC_ID
whichAttr(bp)
#>   [1] "10% salts"                              
#>   [2] "20% salts"                              
#>   [3] "30% salts"                              
#>   [4] "5% salts"                               
#>   [5] "5% sheep blood agar"                    
#>   [6] "Acne"                                   
#>   [7] "Arthritis"                              
#>   [8] "Blunervirus"                            
#>   [9] "Burkholderia cepacia"                   
#>  [10] "Butyrate-Producing Bacteria"            
#>  [11] "CAMP test"                              
#>  [12] "COGEM pathogenicity rating 1"           
#>  [13] "COGEM pathogenicity rating 2"           
#>  [14] "COGEM pathogenicity rating 3"           
#>  [15] "Caries"                                 
#>  [16] "Conjunctivitis"                         
#>  [17] "DSMZ Medium 1000"                       
#>  [18] "Extreme environment"                    
#>  [19] "Flavobacterium psychrophilum"           
#>  [20] "Foodborne Illness"                      
#>  [21] "Furunculosis"                           
#>  [22] "Gangrene"                               
#>  [23] "Gastroenteritis"                        
#>  [24] "Glanders"                               
#>  [25] "Gonorrhea"                              
#>  [26] "Haemophilus"                            
#>  [27] "Health associated"                      
#>  [28] "Infertility"                            
#>  [29] "Inflammation"                           
#>  [30] "Lyme Disease"                           
#>  [31] "Microcystis aeruginosa"                 
#>  [32] "Mycobacterial Infection"                
#>  [33] "Periodontal disorder"                   
#>  [34] "Plant chlorosis"                        
#>  [35] "Reaction to Leaf spot"                  
#>  [36] "Sphingolipid producing"                 
#>  [37] "Terrabacteria group"                    
#>  [38] "Toxic Shock Syndrome"                   
#>  [39] "Tuberculosis"                           
#>  [40] "Tumor"                                  
#>  [41] "ZoBell marine agar"                     
#>  [42] "abortion"                               
#>  [43] "abscess"                                
#>  [44] "acetate producing"                      
#>  [45] "acidic"                                 
#>  [46] "acrodermatitis"                         
#>  [47] "aerobic"                                
#>  [48] "aerotolerant"                           
#>  [49] "african tick-bite fever"                
#>  [50] "alkaline"                               
#>  [51] "alpha"                                  
#>  [52] "anaerobic"                              
#>  [53] "anaplasmosis"                           
#>  [54] "anemia"                                 
#>  [55] "animal pathogen"                        
#>  [56] "anthrax"                                
#>  [57] "antimicrobial sensitivity"              
#>  [58] "artificial seawater"                    
#>  [59] "asp-2"                                  
#>  [60] "atypical pneumonia"                     
#>  [61] "average"                                
#>  [62] "bacillary angiomatosis"                 
#>  [63] "bacillus"                               
#>  [64] "bacteremia"                             
#>  [65] "bacterial leaf blight"                  
#>  [66] "bacterial vaginosis"                    
#>  [67] "bartonellosis"                          
#>  [68] "bent filament"                          
#>  [69] "beta"                                   
#>  [70] "bg11"                                   
#>  [71] "biofilm formation"                      
#>  [72] "biosafety level 1"                      
#>  [73] "biosafety level 2"                      
#>  [74] "biosafety level 3"                      
#>  [75] "blood infection"                        
#>  [76] "bordetellosis"                          
#>  [77] "botulism"                               
#>  [78] "brackish microbiological culture medium"
#>  [79] "branched"                               
#>  [80] "brucellosis"                            
#>  [81] "c"                                      
#>  [82] "canker"                                 
#>  [83] "casitone-yeast extract agar"            
#>  [84] "cat scratch fever"                      
#>  [85] "cell chain"                             
#>  [86] "cell cluster"                           
#>  [87] "cell lysis"                             
#>  [88] "charcoal-yeast extract"                 
#>  [89] "chocolate agar"                         
#>  [90] "cholera"                                
#>  [91] "chronic granulomatous disease"          
#>  [92] "clavate"                                
#>  [93] "coccobacillus"                          
#>  [94] "coccus"                                 
#>  [95] "colitis"                                
#>  [96] "conway"                                 
#>  [97] "crown gall"                             
#>  [98] "cystitis"                               
#>  [99] "defined microbiological culture medium" 
#> [100] "diarrhea"                               
#> [101] "diphtheria"                             
#> [102] "discoid"                                
#> [103] "dumbbell-shaped"                        
#> [104] "duodenal lesions"                       
#> [105] "dysentery"                              
#> [106] "ehrlichiosis"                           
#> [107] "elliptic"                               
#> [108] "encephalitis"                           
#> [109] "endocarditis"                           
#> [110] "endospore"                              
#> [111] "erysipelas"                             
#> [112] "facultatively anaerobic"                
#> [113] "fast"                                   
#> [114] "fever"                                  
#> [115] "filamentous"                            
#> [116] "fildes enrichment"                      
#> [117] "flat"                                   
#> [118] "gamma"                                  
#> [119] "gastroenteritis"                        
#> [120] "gram stain negative"                    
#> [121] "gram stain positive"                    
#> [122] "gram stain variable"                    
#> [123] "gse"                                    
#> [124] "heart infusion agar"                    
#> [125] "heartwater"                             
#> [126] "human granulocytic anaplasmosis"        
#> [127] "hydrogen gas producing"                 
#> [128] "hyperthermophile"                       
#> [129] "infection"                              
#> [130] "kidney disease"                         
#> [131] "lactate producing"                      
#> [132] "large"                                  
#> [133] "lb agar"                                
#> [134] "lb medium"                              
#> [135] "legionnaire's disease"                  
#> [136] "leprosy"                                
#> [137] "mastitis"                               
#> [138] "medium"                                 
#> [139] "medium 31 agar"                         
#> [140] "melioidosis"                            
#> [141] "meningitis"                             
#> [142] "meningoencephalitis"                    
#> [143] "mesophile"                              
#> [144] "modified shieh agar"                    
#> [145] "motility"                               
#> [146] "mrs broth"                              
#> [147] "mueller-hinton"                         
#> [148] "neutral"                                
#> [149] "nutrient broth"                         
#> [150] "paired cells"                           
#> [151] "palisade"                               
#> [152] "pathogenicity human"                    
#> [153] "plague"                                 
#> [154] "plant pathogenity"                      
#> [155] "pleiomorphic cell"                      
#> [156] "pneumonia"                              
#> [157] "polyarthritis"                          
#> [158] "polyhedral shape"                       
#> [159] "psychrophile"                           
#> [160] "pyelonephritis"                         
#> [161] "pyg medium"                             
#> [162] "reaction to black rot"                  
#> [163] "rectangle"                              
#> [164] "respiratory infection"                  
#> [165] "respiratory system disease"             
#> [166] "reverse Y-branched filaments"           
#> [167] "rosettes"                               
#> [168] "septicemia"                             
#> [169] "sigmoid"                                
#> [170] "single"                                 
#> [171] "skin irritation"                        
#> [172] "slow"                                   
#> [173] "sm agar"                                
#> [174] "small"                                  
#> [175] "sorghum"                                
#> [176] "spirillum"                              
#> [177] "spore formation"                        
#> [178] "staley pyg"                             
#> [179] "staphylococcus"                         
#> [180] "syphilis"                               
#> [181] "tail"                                   
#> [182] "tetrads"                                
#> [183] "tetrahedral"                            
#> [184] "thermophile"                            
#> [185] "toxin"                                  
#> [186] "trench fever"                           
#> [187] "triangle"                               
#> [188] "tryptic soy agar"                       
#> [189] "tryptic soy broth"                      
#> [190] "tuberculosis"                           
#> [191] "tumors"                                 
#> [192] "typhoid"                                
#> [193] "typhus"                                 
#> [194] "u-shaped"                               
#> [195] "ulcer"                                  
#> [196] "unbranched"                             
#> [197] "urinary tract infection"                
#> [198] "v-formation"                            
#> [199] "very alkaline"                          
#> [200] "very large"                             
#> [201] "very small"                             
#> [202] "wrinkled"