Which Attributes

whichAttr shows which attributes are present in a dataset imported with importBugphyzz. This would be the names of the singatures created with getBugphyzzSignatures.

whichAttr(bp)

Arguments

bp

A data.frame imported with importBugphyzz.

Value

A character vector.

Examples

bp <- importBugphyzz()
#> Using data downloaded on 2023-07-17 14:21:22.
#> Warning: The following named parsers don't match the column names: BacDive_ID, Type_strain, PATRIC_ID
whichAttr(bp)
#>   [1] "Acne"                            "Arthritis"                      
#>   [3] "Blunervirus"                     "Burkholderia cepacia"           
#>   [5] "Butyrate-Producing Bacteria"     "COGEM pathogenicity rating 1"   
#>   [7] "COGEM pathogenicity rating 2"    "COGEM pathogenicity rating 3"   
#>   [9] "Caries"                          "Conjunctivitis"                 
#>  [11] "Extreme environment"             "Flavobacterium psychrophilum"   
#>  [13] "Foodborne Illness"               "Furunculosis"                   
#>  [15] "Gangrene"                        "Gastroenteritis"                
#>  [17] "Glanders"                        "Gonorrhea"                      
#>  [19] "Haemophilus"                     "Health associated"              
#>  [21] "Infertility"                     "Inflammation"                   
#>  [23] "Lyme Disease"                    "Microcystis aeruginosa"         
#>  [25] "Mycobacterial Infection"         "Periodontal disorder"           
#>  [27] "Plant chlorosis"                 "Reaction to Leaf spot"          
#>  [29] "Sphingolipid producing"          "Terrabacteria group"            
#>  [31] "Toxic Shock Syndrome"            "Tuberculosis"                   
#>  [33] "Tumor"                           "abortion"                       
#>  [35] "abscess"                         "acetate producing"              
#>  [37] "acidic"                          "acrodermatitis"                 
#>  [39] "african tick-bite fever"         "alkaline"                       
#>  [41] "anaplasmosis"                    "anemia"                         
#>  [43] "anthrax"                         "antimicrobial sensitivity"      
#>  [45] "atypical pneumonia"              "average"                        
#>  [47] "bacillary angiomatosis"          "bacillus"                       
#>  [49] "bacteremia"                      "bacterial leaf blight"          
#>  [51] "bacterial vaginosis"             "bartonellosis"                  
#>  [53] "biofilm formation"               "blood infection"                
#>  [55] "bordetellosis"                   "botulism"                       
#>  [57] "branched"                        "brucellosis"                    
#>  [59] "canker"                          "cat scratch fever"              
#>  [61] "cell chain"                      "cell cluster"                   
#>  [63] "cell lysis"                      "cholera"                        
#>  [65] "chronic granulomatous disease"   "coccus"                         
#>  [67] "colitis"                         "crown gall"                     
#>  [69] "cystitis"                        "diarrhea"                       
#>  [71] "diphtheria"                      "duodenal lesions"               
#>  [73] "dysentery"                       "ehrlichiosis"                   
#>  [75] "elliptic"                        "encephalitis"                   
#>  [77] "endocarditis"                    "endospore"                      
#>  [79] "erysipelas"                      "fast"                           
#>  [81] "fever"                           "filamentous"                    
#>  [83] "gastroenteritis"                 "heartwater"                     
#>  [85] "human granulocytic anaplasmosis" "hydrogen gas producing"         
#>  [87] "infection"                       "kidney disease"                 
#>  [89] "lactate producing"               "large"                          
#>  [91] "legionnaire's disease"           "leprosy"                        
#>  [93] "mastitis"                        "medium"                         
#>  [95] "melioidosis"                     "meningitis"                     
#>  [97] "meningoencephalitis"             "neutral"                        
#>  [99] "paired cells"                    "palisade"                       
#> [101] "plague"                          "plant pathogenity"              
#> [103] "pneumonia"                       "polyarthritis"                  
#> [105] "pyelonephritis"                  "reaction to black rot"          
#> [107] "respiratory infection"           "respiratory system disease"     
#> [109] "reverse Y-branched filaments"    "rosettes"                       
#> [111] "septicemia"                      "sigmoid"                        
#> [113] "single"                          "skin irritation"                
#> [115] "slow"                            "small"                          
#> [117] "sorghum"                         "spirillum"                      
#> [119] "syphilis"                        "tetrads"                        
#> [121] "toxin"                           "trench fever"                   
#> [123] "tuberculosis"                    "tumors"                         
#> [125] "typhoid"                         "typhus"                         
#> [127] "ulcer"                           "unbranched"                     
#> [129] "urinary tract infection"         "v-formation"                    
#> [131] "very alkaline"                   "very large"                     
#> [133] "very small"