Which Attributes

whichAttr shows which attributes are present in a dataset imported with importBugphyzz. This would be the names of the singatures created with getBugphyzzSignatures.

whichAttr(bp)

Arguments

bp

A data.frame imported with importBugphyzz.

Value

A character vector.

Examples

bp <- importBugphyzz()
#> Using data downloaded on 2023-07-24 13:34:40.
#> Warning: The following named parsers don't match the column names: Strain, PATRIC_ID
whichAttr(bp)
#>   [1] "10% salts"                              
#>   [2] "20% salts"                              
#>   [3] "30% salts"                              
#>   [4] "5% salts"                               
#>   [5] "5% sheep blood agar"                    
#>   [6] "Acne"                                   
#>   [7] "Arthritis"                              
#>   [8] "Blunervirus"                            
#>   [9] "Burkholderia cepacia"                   
#>  [10] "Butyrate-Producing Bacteria--TRUE"      
#>  [11] "CAMP test"                              
#>  [12] "COGEM pathogenicity rating 1"           
#>  [13] "COGEM pathogenicity rating 2"           
#>  [14] "COGEM pathogenicity rating 3"           
#>  [15] "Caries"                                 
#>  [16] "Conjunctivitis"                         
#>  [17] "DSMZ Medium 1000"                       
#>  [18] "Extreme environment--FALSE"             
#>  [19] "Extreme environment--TRUE"              
#>  [20] "Flavobacterium psychrophilum"           
#>  [21] "Foodborne Illness"                      
#>  [22] "Furunculosis"                           
#>  [23] "Gangrene"                               
#>  [24] "Gastroenteritis"                        
#>  [25] "Glanders"                               
#>  [26] "Gonorrhea"                              
#>  [27] "Haemophilus"                            
#>  [28] "Health associated--FALSE"               
#>  [29] "Health associated--TRUE"                
#>  [30] "Infertility"                            
#>  [31] "Inflammation"                           
#>  [32] "Lyme Disease"                           
#>  [33] "Microcystis aeruginosa"                 
#>  [34] "Mycobacterial Infection"                
#>  [35] "Periodontal disorder"                   
#>  [36] "Plant chlorosis"                        
#>  [37] "Reaction to Leaf spot"                  
#>  [38] "Sphingolipid producing--TRUE"           
#>  [39] "Terrabacteria group"                    
#>  [40] "Toxic Shock Syndrome"                   
#>  [41] "Tuberculosis"                           
#>  [42] "Tumor"                                  
#>  [43] "ZoBell marine agar"                     
#>  [44] "abortion"                               
#>  [45] "abscess"                                
#>  [46] "acetate producing--TRUE"                
#>  [47] "acidic"                                 
#>  [48] "acrodermatitis"                         
#>  [49] "aerobic"                                
#>  [50] "aerotolerant"                           
#>  [51] "african tick-bite fever"                
#>  [52] "alkaline"                               
#>  [53] "alpha"                                  
#>  [54] "anaerobic"                              
#>  [55] "anaplasmosis"                           
#>  [56] "anemia"                                 
#>  [57] "animal pathogen--FALSE"                 
#>  [58] "animal pathogen--TRUE"                  
#>  [59] "anthrax"                                
#>  [60] "antimicrobial sensitivity--FALSE"       
#>  [61] "antimicrobial sensitivity--TRUE"        
#>  [62] "artificial seawater"                    
#>  [63] "asp-2"                                  
#>  [64] "atypical pneumonia"                     
#>  [65] "average"                                
#>  [66] "bacillary angiomatosis"                 
#>  [67] "bacillus"                               
#>  [68] "bacteremia"                             
#>  [69] "bacterial leaf blight"                  
#>  [70] "bacterial vaginosis"                    
#>  [71] "bartonellosis"                          
#>  [72] "bent filament"                          
#>  [73] "beta"                                   
#>  [74] "bg11"                                   
#>  [75] "biofilm formation--FALSE"               
#>  [76] "biofilm formation--TRUE"                
#>  [77] "biosafety level 1"                      
#>  [78] "biosafety level 2"                      
#>  [79] "biosafety level 3"                      
#>  [80] "blood infection"                        
#>  [81] "bordetellosis"                          
#>  [82] "botulism"                               
#>  [83] "brackish microbiological culture medium"
#>  [84] "branched"                               
#>  [85] "brucellosis"                            
#>  [86] "c"                                      
#>  [87] "canker"                                 
#>  [88] "casitone-yeast extract agar"            
#>  [89] "cat scratch fever"                      
#>  [90] "cell chain"                             
#>  [91] "cell cluster"                           
#>  [92] "cell lysis"                             
#>  [93] "charcoal-yeast extract"                 
#>  [94] "chocolate agar"                         
#>  [95] "cholera"                                
#>  [96] "chronic granulomatous disease"          
#>  [97] "clavate"                                
#>  [98] "coccobacillus"                          
#>  [99] "coccus"                                 
#> [100] "colitis"                                
#> [101] "conway"                                 
#> [102] "crown gall"                             
#> [103] "cystitis"                               
#> [104] "defined microbiological culture medium" 
#> [105] "diarrhea"                               
#> [106] "diphtheria"                             
#> [107] "discoid"                                
#> [108] "dumbbell-shaped"                        
#> [109] "duodenal lesions"                       
#> [110] "dysentery"                              
#> [111] "ehrlichiosis"                           
#> [112] "elliptic"                               
#> [113] "encephalitis"                           
#> [114] "endocarditis"                           
#> [115] "endospore"                              
#> [116] "erysipelas"                             
#> [117] "facultatively anaerobic"                
#> [118] "fast"                                   
#> [119] "fever"                                  
#> [120] "filamentous"                            
#> [121] "fildes enrichment"                      
#> [122] "flat"                                   
#> [123] "gamma"                                  
#> [124] "gastroenteritis"                        
#> [125] "gram stain negative"                    
#> [126] "gram stain positive"                    
#> [127] "gram stain variable"                    
#> [128] "gse"                                    
#> [129] "heart infusion agar"                    
#> [130] "heartwater"                             
#> [131] "human granulocytic anaplasmosis"        
#> [132] "hydrogen gas producing--FALSE"          
#> [133] "hydrogen gas producing--TRUE"           
#> [134] "hyperthermophile"                       
#> [135] "infection"                              
#> [136] "kidney disease"                         
#> [137] "lactate producing--FALSE"               
#> [138] "lactate producing--TRUE"                
#> [139] "large"                                  
#> [140] "lb agar"                                
#> [141] "lb medium"                              
#> [142] "legionnaire's disease"                  
#> [143] "leprosy"                                
#> [144] "mastitis"                               
#> [145] "medium"                                 
#> [146] "medium 31 agar"                         
#> [147] "melioidosis"                            
#> [148] "meningitis"                             
#> [149] "meningoencephalitis"                    
#> [150] "mesophile"                              
#> [151] "modified shieh agar"                    
#> [152] "motility--FALSE"                        
#> [153] "motility--TRUE"                         
#> [154] "mrs broth"                              
#> [155] "mueller-hinton"                         
#> [156] "neutral"                                
#> [157] "nutrient broth"                         
#> [158] "paired cells"                           
#> [159] "palisade"                               
#> [160] "pathogenicity human--TRUE"              
#> [161] "plague"                                 
#> [162] "plant pathogenity--FALSE"               
#> [163] "plant pathogenity--TRUE"                
#> [164] "pleiomorphic cell"                      
#> [165] "pneumonia"                              
#> [166] "polyarthritis"                          
#> [167] "polyhedral shape"                       
#> [168] "psychrophile"                           
#> [169] "pyelonephritis"                         
#> [170] "pyg medium"                             
#> [171] "reaction to black rot"                  
#> [172] "rectangle"                              
#> [173] "respiratory infection"                  
#> [174] "respiratory system disease"             
#> [175] "reverse Y-branched filaments"           
#> [176] "rosettes"                               
#> [177] "septicemia"                             
#> [178] "sigmoid"                                
#> [179] "single"                                 
#> [180] "skin irritation"                        
#> [181] "slow"                                   
#> [182] "sm agar"                                
#> [183] "small"                                  
#> [184] "sorghum"                                
#> [185] "spirillum"                              
#> [186] "spore formation--FALSE"                 
#> [187] "spore formation--TRUE"                  
#> [188] "staley pyg"                             
#> [189] "staphylococcus"                         
#> [190] "syphilis"                               
#> [191] "tail"                                   
#> [192] "tetrads"                                
#> [193] "tetrahedral"                            
#> [194] "thermophile"                            
#> [195] "toxin"                                  
#> [196] "trench fever"                           
#> [197] "triangle"                               
#> [198] "tryptic soy agar"                       
#> [199] "tryptic soy broth"                      
#> [200] "tuberculosis"                           
#> [201] "tumors"                                 
#> [202] "typhoid"                                
#> [203] "typhus"                                 
#> [204] "u-shaped"                               
#> [205] "ulcer"                                  
#> [206] "unbranched"                             
#> [207] "urinary tract infection"                
#> [208] "v-formation"                            
#> [209] "very alkaline"                          
#> [210] "very large"                             
#> [211] "very small"                             
#> [212] "wrinkled"