Which Attributes

whichAttr shows which attributes are present in a dataset imported with importBugphyzz. This would be the names of the singatures created with getBugphyzzSignatures.

whichAttr(bp)

Arguments

bp

A data.frame imported with importBugphyzz.

Value

A character vector.

Examples

bp <- importBugphyzz()
#> Using data downloaded on 2023-05-19 16:36:25.
whichAttr(bp)
#>  [1] "aerophilicity:aerobic"                
#>  [2] "aerophilicity:aerotolerant"           
#>  [3] "aerophilicity:anaerobic"              
#>  [4] "aerophilicity:facultatively_anaerobic"
#>  [5] "aerophilicity:microaerophilic"        
#>  [6] "aerophilicity:obligately_aerobic"     
#>  [7] "aerophilicity:obligately_anaerobic"   
#>  [8] "gram_stain:gram_stain_negative"       
#>  [9] "gram_stain:gram_stain_positive"       
#> [10] "gram_stain:gram_stain_variable"       
#> [11] "growth_temperature:hyperthermophile"  
#> [12] "growth_temperature:mesophile"         
#> [13] "growth_temperature:psychrophile"      
#> [14] "growth_temperature:thermophile"       
#> [15] "optimal_ph:acidic"                    
#> [16] "optimal_ph:alkaline"                  
#> [17] "optimal_ph:neutral"                   
#> [18] "optimal_ph:very_alkaline"             
#> [19] "shape:bacillus"                       
#> [20] "shape:bent_filament"                  
#> [21] "shape:clavate"                        
#> [22] "shape:coccobacillus"                  
#> [23] "shape:coccus"                         
#> [24] "shape:discoid"                        
#> [25] "shape:dumbbell-shaped"                
#> [26] "shape:elliptic"                       
#> [27] "shape:filamentous"                    
#> [28] "shape:flat"                           
#> [29] "shape:pleiomorphic_cell"              
#> [30] "shape:polyhedral_shape"               
#> [31] "shape:rectangle"                      
#> [32] "shape:sigmoid"                        
#> [33] "shape:spirillum"                      
#> [34] "shape:staphylococcus"                 
#> [35] "shape:tail"                           
#> [36] "shape:tetrahedral"                    
#> [37] "shape:triangle"                       
#> [38] "shape:u-shaped"                       
#> [39] "shape:wrinkled"                       
#> [40] "width:average"                        
#> [41] "width:large"                          
#> [42] "width:small"                          
#> [43] "width:very_large"